Refactor hello world submodule with path
