+++ src/app/app.component.html
@@ -2,7 +2,6 @@
   <span>BeeFinder</span>
 </mat-toolbar>
 <main>
-  <app-hello-world></app-hello-world>
   <router-outlet></router-outlet>
 </main>
 <mat-toolbar >
