+++ src/app/app.module.ts
@@ -5,10 +5,12 @@ import { NgModule } from '@angular/core';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
 import { environment } from '../environments/environment';
-import { HelloWorldModule } from './hello-world/hello-world.module';
+import { HelloWorldModule } from './views/hello-world/hello-world.module';
 import { AngularFirestoreModule } from '@angular/fire/firestore';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { ComponentsModule } from './components/components.module';
+import { RouterModule } from '@angular/router';
+import { appRoutes } from 'src/routes';
 
 @NgModule({
   declarations: [
@@ -17,6 +19,9 @@ import { ComponentsModule } from './components/components.module';
   imports: [
     BrowserModule,
     AppRoutingModule,
+    RouterModule.forRoot(
+      appRoutes
+    ),
     AngularFireModule.initializeApp(environment.firebase),
     AngularFireAuthModule,
     AngularFirestoreModule,
