+++ src/routes.ts
@@ -1,0 +1,6 @@
+import { Routes } from '@angular/router';
+import { HelloWorldComponent } from './app/views/hello-world/components/messages.component';
+
+export const appRoutes: Routes = [
+  { path: 'hello-world', component: HelloWorldComponent }
+];
